[#176] CI build fails due to not able to download dependencies
CI build is failing intermittently while downloading the GAE dev server
dependencies.

We cannot mount the /root/.m2 repository because the image has there the
maven binaries.

Changing the maven config to use a non-standard dir for the local repo,
so that we can mount that path and let Travis cache the dependencies.