[#172] Flow to create the report
Instead of flow-services creating the report, the Flow backend will
proxy the UI requests.

Flow will send exactly the same request than the UI but it will add the
id of the report in the flow side.

Flow-services will still notify Flow when starting generating the report
and once it is finished