[#219] loading settings and config
Settings and config was nil because we are running the test from the
"test" container, which doesnt do any initialization.

We now conditionally init the settings, so that we dont change the
config when working on the repl.

Passing the settings and config as parameters is too much work as all
functions are willy nilly referring to them.