[#265] Update Flow dependency + Invalidate cache between tests
I have tried to copy a "sample" file and use System/currentTimemillis,
but somehow the Java code that opens the connection and reads the XML
was failing to parse the copy with a very cryptic error:

    Unexpected EOF in prolog at [row,col {unknown-source}]: [1,0]

It suggests that the file is corrupted or something similar. When
verifying the integrity of the copy (with `sha256sum`) the
resulting file was "identical".

As @stellanl suggested we now issue a POST request to `/invalidate` to
avoid using a cached report between tests.